platform/api/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.4